feat : 임의의 버스 현재 위치 그리기 구현

+++ data/src/main/java/com/stop/data/model/nowlocation/BusInfoRepositoryItem.kt
@@ -14,8 +14,8 @@ data class BusInfoRepositoryItem(
     fun toUseCaseModel() = BusInfoUseCaseItem(
         isArrivedAtStation,
         sectionId,
-        latitude,
-        longitude,
+        latitude.toDouble(),
+        longitude.toDouble(),
         busId,
         busNumber,
         isRun,
